Do not panic when regex doesn't match
This can happen when # is converted to * but matches a different file
than the numbered file. For example if we have files like

box_1.vtk
box_2.vtk
...
box_other.vtk

then the pattern box_#.vtk will match all of them, but "box_other.vtk"
will cause the program to panic. This commit fixes this issue.