Remove orphaned vertices
This change removes vertices on triangle meshes that also contain tets.
Previously vertices belonging to the tetmesh were included in the
triangle mesh and ultimately duplicated. While this is ignored in most
applications, when writing out attributes (e.g. pressure), the extra
vertices can be problematic (e.g. the tri_and_tet example fails to load
in Blender). This commit is one step towards resolving this issue,
although there is a problem with sharing position data between multiple
primitives which should be resolved in a latter commit.